+++ changelog.md
@@ -6,7 +6,9 @@ SchemaCrawler AI release notes.
 <a name="v17.11.4-3"></a>
 ## Release 17.11.4-3 - 2026-06-30
 
+- Rename database server information function
 - Return primary keys in table description by default
+- Show next steps for each tool call
 
 
 <a name="v17.11.4-2"></a>